+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-19T09:07:33.562Z",
+    "2026-02-19T11:04:28.599Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {